backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pdfplumber
python-multipart
networkx
